# backend-express/docs/openapi/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Drivee Link API (Express)",
    "version": "1.0.0",
    "description": "OpenAPI path keys **không** có tiền tố `/api` vì `servers.url` trong `config/swagger.js` đã là `.../api`.\n\nKhai báo operation trong controller bằng JSDoc `@openapi` (đọc bởi `swagger-jsdoc`). Map từ Laravel: bỏ `/api` ở đầu path.\n",
    "contact": {
      "name": "ATMC Development Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3400/api",
      "description": "production"
    }
  ],
...
  "components": {
...
    "securitySchemes": {
      "auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token trả về từ POST /auth/login (field access_token, có thể đã gồm tiền tố \"Bearer \")."
      }
    }
  },
  "paths": {
...
    "/auth/forget-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forget password",
        "operationId": "forget_password",
        "description": "Laravel `@OA` dùng query `mail_address`; Express dùng JSON body.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mail_address"
                ],
                "properties": {
                  "mail_address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gửi yêu cầu thành công",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "data": "Password reset URL has been sent to your email address"
                }
              }
            }
          }
        }
      }
    },
...
  }
}